support multiple CRN_Model versions
