+++ .gitignore
@@ -1,0 +1,2 @@
+results
+*__pycache__/
